Test piece first impl
